pyproject.toml
[project]
name = "news-search-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.12",
    "multipledispatch>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "s3p-sdk>=0.2.6",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.2",
]
